feat: implementation of controllers

+++ src/use-cases/delete-user.ts
@@ -1,8 +1,9 @@
 import { User } from "../@types/user"
 import { UsersRepository } from "../http/repositories/users-repository"
+import { ResourceNotFoundError } from "./errors/resource-not-found-error"
 
 interface DeleteUserUseCaseRequest {
-    user: User
+    id: string
 }
 
 interface DeleteUserUseCaseResponse {
@@ -13,9 +14,16 @@ export class DeleteUserUseCase {
     constructor (private usersRepository: UsersRepository){}
 
     async execute({ 
-        user
+        id
 
      }: DeleteUserUseCaseRequest): Promise<DeleteUserUseCaseResponse> {
+
+        const user = await this.usersRepository.findById(id)
+
+        if (!user) {
+            throw new ResourceNotFoundError()
+        }
+
         await this.usersRepository.delete(user)
 
         return {
